Editor Package refactor
