+++ src/main/scala/view/form/FormElement.scala
@@ -1,10 +1,10 @@
-package view.editor
+package view.form
 
 import view.util.scalaQuestSwingComponents.{SqSwingBoxPanel, SqSwingLabel}
 
 import java.text.NumberFormat
+import javax.swing._
 import javax.swing.text.NumberFormatter
-import javax.swing.{BoxLayout, JFormattedTextField, JScrollPane, JTextArea, JTextField}
 
 abstract class FormElement(textLabel: String) extends SqSwingBoxPanel(BoxLayout.Y_AXIS) {
